# 🛠️ Tool 4: HealthCalculator - Algorithmic Scoring and Metrics