## Summary

Implements issue #111 - Complete implementation of HealthCalculator tool for deterministic project health scoring, metrics tracking, and risk assessment.

### Features Implemented

#### ✅ Project Health Scoring
- Overall health score calculation (0.0-1.0 scale) with automatic categorization
- Component-based scoring with configurable weights:
  - Timeline: 30% (progress toward deadlines)
  - Activity: 25% (development activity level)
  - Blockers: 25% (impact of active blockers)
  - Dependencies: 20% (health of dependent projects)
- Weighted score aggregation with detailed breakdown

#### ✅ Timeline Analysis
- `calculate_timeline_progress()`: Milestone completion tracking
- Progress ratio calculation with date-based projections
- On-track determination with 5% tolerance threshold
- Handles missing data gracefully

#### ✅ Activity Scoring
- `calculate_activity_score()`: GitHub activity analysis
- Commit, PR, and issue closure tracking
- Baseline comparison with configurable thresholds
- Activity ratio scoring (high/normal/low/very_low)

#### ✅ Blocker Impact Assessment
- `calculate_blocker_impact()`: Blocker count and severity analysis
- Severity breakdown tracking
- Inverse scoring (fewer blockers = higher score)
- Impact thresholds: none/minor/moderate/significant/critical

#### ✅ Team Performance Metrics
- `calculate_team_performance()`: Multi-dimensional team metrics
- Velocity, quality, collaboration, throughput, cycle time
- Configurable metric selection
- Structured TeamPerformanceMetrics output

#### ✅ Trend Analysis
- `analyze_trends()`: Historical data analysis with linear regression
- Direction detection: improving/stable/declining
- Confidence scoring based on variance
- Prediction of next values
- Configurable time windows (default: 30 days)

#### ✅ Completion Prediction
- `predict_completion()`: ETA calculation based on velocity
- Risk level assessment (low/medium/high)
- Confidence scoring
- Handles edge cases (zero velocity, etc.)

#### ✅ Risk Assessment
- `assess_risks()`: Comprehensive project risk identification
- Timeline delay detection with severity classification
- High blocker count alerts
- Low activity warnings
- Priority scoring: (severity × 0.6) + (likelihood × 0.4)
- Mitigation suggestions for each risk

#### ✅ Configuration System
- Default weights following specification
- Customizable thresholds for all components
- Health category boundaries
- Risk severity and likelihood mappings
- Trend analysis settings
- Weight validation (must sum to 1.0)

#### ✅ Data Models
- HealthScore: Complete health breakdown
- ComponentScore: Individual component details
- TrendAnalysis: Trend data with predictions
- Risk: Risk identification with mitigation
- TeamPerformanceMetrics: Team performance tracking
- Comprehensive enum types for categories

## Test Coverage

### Test Statistics
- **Total tests**: 37 tests, all passing ✅
- **Coverage**:
  - `health_calculator.py`: 95% coverage
  - `health_config.py`: 86% coverage
  - `health_models.py`: 97% coverage
  - **Overall**: >80% target exceeded ✅

### Test Categories
- **Project Health**: 4 tests (success, custom weights, empty data, performance)
- **Timeline Progress**: 3 tests (milestones, dates, empty)
- **Activity Scoring**: 4 tests (high/low/baseline/empty)
- **Blocker Impact**: 2 tests (none/multiple)
- **Team Performance**: 2 tests (full/defaults)
- **Trend Analysis**: 5 tests (improving/declining/stable/insufficient/empty)
- **Completion Prediction**: 2 tests (valid/zero velocity)
- **Risk Assessment**: 4 tests (timeline/blockers/activity/sorting)
- **Configuration**: 4 tests (default/custom/invalid/categorization)
- **Models**: 5 tests (score creation/categorization/dict conversion/priority)
- **Performance**: 2 tests (stats tracking/no calculations)

## Performance Benchmarks

✅ All calculations complete within **<50ms** target:
- Average: ~10ms per calculation (5x faster than spec)
- Performance tracking built-in
- Deterministic results for testing
- Efficient caching support

## API Simplification

**Before** (80+ lines of manual calculations):
```python
# Manual health score calculation with complex logic
timeline_score = ...  # 15 lines
activity_score = ...  # 20 lines  
blocker_score = ...   # 15 lines
dependency_score = ... # 10 lines
overall = (timeline * 0.3 + activity * 0.25 + ...) # 20+ lines
```

**After** (1-3 lines):
```python
calc = HealthCalculator()
health = calc.calculate_project_health(project_data)
print(f"{health.category.value}: {health.score:.2f}")
```

## Integration Points

### Ready for Command Integration
- `/project-status`: Replace manual health calculations
- `/project-review`: Add trend analysis
- `/project-risks`: Use risk assessment

### Tool Compatibility
- Works seamlessly with DataCollector
- Compatible with existing project data structures
- All classes exported via `tools/__init__.py`

## File Structure

```
tools/
├── health_calculator.py    # Main calculator (165 statements, 95% coverage)
├── health_models.py        # Data models (106 statements, 97% coverage)
├── health_config.py        # Configuration (81 statements, 86% coverage)
└── __init__.py            # Updated exports

tests/
└── test_health_calculator.py  # 37 comprehensive tests
```

## Usage Examples

### Basic Health Calculation
```python
from tools import HealthCalculator

calc = HealthCalculator()
health = calc.calculate_project_health({
    "milestones": [...],
    "github_data": {...},
    "blockers": [...],
    "dependencies": [...]
})

print(f"Health: {health.score:.2f} ({health.category.value})")
# Output: Health: 0.72 (good)

for name, component in health.components.items():
    print(f"{name}: {component.score:.2f}")
# Output:
# timeline: 0.80
# activity: 0.70
# blockers: 0.75
# dependencies: 0.85
```

### Trend Analysis
```python
trends = calc.analyze_trends([
    {"timestamp": "2025-09-01", "value": 0.60},
    {"timestamp": "2025-09-15", "value": 0.70},
    {"timestamp": "2025-09-29", "value": 0.80},
])

print(f"Direction: {trends.direction.value}")
print(f"Confidence: {trends.confidence:.2f}")
# Output:
# Direction: improving
# Confidence: 0.87
```

### Risk Assessment
```python
risks = calc.assess_risks(project_data)
for risk in risks:
    print(f"{risk.severity.value}: {risk.title}")
    print(f"Priority: {risk.priority_score:.2f}")
# Output sorted by priority score
```

### Custom Configuration
```python
custom_config = HealthConfig({
    "weights": {
        "timeline": 0.4,
        "activity": 0.3,
        "blockers": 0.2,
        "dependencies": 0.1,
    }
})

calc = HealthCalculator(config=custom_config)
health = calc.calculate_project_health(project_data)
```

## Success Criteria

✅ All scoring algorithms implemented  
✅ Test coverage >80% (achieved 86-97%)  
✅ Documentation complete with formulas  
✅ Performance <50ms (achieved ~10ms average)  
✅ Results match manual calculations  
✅ Configurable weights and thresholds  
✅ Comprehensive error handling  
✅ Type hints throughout  
✅ All edge cases handled

## Next Steps

After merge:
1. Update `/project-status` command to use HealthCalculator
2. Integrate with `/project-review` for trend analysis
3. Update `/project-risks` to use risk assessment
4. Add command integration examples to documentation
5. Consider adding visualization tools for health trends

## Related Issues

- Closes #111
- Depends on #109 (DataCollector) ✅ Already merged

---

🤖 Generated with [Claude Code](https://claude.com/claude-code)